% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineer.R
\name{engineer}
\alias{engineer}
\title{Engineer a Monster combination}
\usage{
engineer(parent, child)
}
\arguments{
\item{parent}{A \code{\link{Monster}} object}

\item{child}{A \code{\link{Monster}} object}
}
\value{
An object of class "MonsterComboEngineering"; a list with elements
\describe{
\item{given_parent}{The \code{\link{Monster}} given in \code{parent}}
\item{given_child}{The \code{\link{Monster}} given in \code{child}}
\item{stat_orders}{A data frame listing potential parents' stat orders}
}
}
\description{
Engineer a Monster combination
}
