% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monster.R
\name{Monster}
\alias{Monster}
\title{Create a Monster object}
\usage{
Monster(name = "", monster = "", main = "", sub = "", stats = NULL)
}
\arguments{
\item{name}{A character vector of length one; the monster's given name}

\item{monster}{A character vector of length one; the monster's breed name;
if \code{""} (the default), this will be determined by the \code{main}
and \code{sub} arguments}

\item{main}{A character vector of length one; the monster's main breed;
if \code{""} (the default), this will be determined by the \code{monster}
argument}

\item{sub}{A character vector of length one; the monster's sub breed;
if \code{""} (the default), this will be determined by the \code{monster}
argument}

\item{stats}{A numeric vector of length six; the monster's current stats;
if \code{NULL} (the default), the baseline stats of the monster breed
will be used; the vector should either be in the order
life, power, intelligence, skill, speed, defense, or else the elements
should have those names}
}
\value{
An object of class "Monster"; a list with elements
\describe{
\item{name}{The monster's given name}
\item{monster}{The monster's breed name}
\item{main}{The monster's main breed}
\item{sub}{The monster's sub breed}
\item{stats}{The monster's current stats}
\item{baselines}{The monster's breed's baseline stats}
\item{statgains}{The monster's breed's statgain information}
\item{lifespan}{The monster's breed's base lifespan (in weeks)}
}
}
\description{
Create a Monster object
}
