% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsterlist}
\alias{monsterlist}
\title{Monster List}
\format{
\subsection{\code{monsterlist}}{

A data frame with 337 rows and 3 columns:
\describe{
\item{monster}{The name of this monster breed}
\item{main}{The main breed of this monster}
\item{sub}{The sub breed of this monster}
}
}
}
\usage{
monsterlist
}
\description{
A list of the Monster Rancher 2 monsters the \code{ranchr} package has data on
}
\keyword{datasets}
