% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monsters}
\alias{monsters}
\title{Monster Rancher 2 Breeding Data}
\format{
\subsection{\code{monsters}}{

A list of length 337, where each element is a \code{\link{Monster}} object
}
}
\usage{
monsters
}
\description{
A list of \code{\link{Monster}} objects,
giving data relevant for monster breeding
}
\keyword{datasets}
