% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{+.Monster}
\alias{+.Monster}
\title{See how two Monsters will combine}
\usage{
\method{+}{Monster}(e1, e2)
}
\arguments{
\item{e1}{A \code{\link{Monster}} object}

\item{e2}{A \code{\link{Monster}} object}
}
\value{
An object of class "MonsterCombination"; a list with elements
\describe{
\item{parent1_stats}{The ordered, adjusted stats of \code{e1}}
\item{parent2_stats}{The ordered, adjusted stats of \code{e2}}
\item{parent_stat_orders}{A data frame listing the parents' stat orders}
\item{baby_stat_orders}{A data frame listing potential babies' stat orders}
\item{best_match}{The predicted best outcome from combining}
}
}
\description{
See how two Monsters will combine
}
